+++ tests/utils.h
@@ -1,5 +1,5 @@
 #include <stdint.h>
 
-void dbg_buff_print(uint8_t *buff, int count);
+void dbg_buff_print(uint8_t *buff, uint64_t count);
 
-bool compareByteArray(uint8_t *first, uint8_t *second, int count);
+bool compareByteArray(uint8_t *first, uint8_t *second, uint64_t count);
